kshop/requirements.txt
Babel==2.6.0
Django==2.2
django-extensions==2.1.6
django-filter==2.1.0
django-grappelli==2.12.2
django-phonenumber-field==2.4.0
djangorestframework==3.9.2
drf-extensions==0.4.0
phonenumbers==8.10.11
Pillow==6.0.0
psycopg2-binary==2.8.2
python-decouple==3.1
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
